+++ commands/export.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"flag"
@@ -6,44 +6,43 @@ import (
 	"os"
 
 	"github.com/transcriptaze/midiasm/midi"
-	"github.com/transcriptaze/midiasm/ops/export"
+	impl "github.com/transcriptaze/midiasm/ops/export"
 )
 
-type Export struct {
+type export struct {
 	out string
 }
 
-var EXPORT = Export{}
+var Export = export{}
 
-func (x *Export) Flagset(flagset *flag.FlagSet) *flag.FlagSet {
+func (x *export) Flagset(flagset *flag.FlagSet) *flag.FlagSet {
 	flagset.StringVar(&x.out, "out", "", "Output file path (or directory for split files)")
 
 	return flagset
 }
 
-func (x Export) Help() {
+func (x export) Help() {
 	fmt.Println()
 	fmt.Println("  Extracts the MIDI information as JSON for use with other tools (e.g. jq).")
 	fmt.Println()
 	fmt.Println("    midiasm export [--debug] [--verbose] [--C4] [--out <file>] <MIDI file>")
 	fmt.Println()
-	fmt.Println("      --out <file>     Writes the JSON to a file. Default is to write to stdout.")
-	fmt.Println("      --json           Formats the output as JSON - the default is human readable text.")
-	fmt.Println("      --transpose <N>  Transposes the notes up or down by N semitones.")
+	fmt.Println("      <MIDI file>  MIDI file to export as JSON.")
 	fmt.Println()
 	fmt.Println("    Options:")
 	fmt.Println()
-	fmt.Println("      --C4       Uses C4 as middle C (Yamaha convention). Defaults to C3.")
-	fmt.Println("      --debug    Displays internal information while processing a MIDI file. Defaults to false")
-	fmt.Println("      --verbose  Enables 'verbose' logging. Defaults to false")
+	fmt.Println("      --out <file>  Writes the JSON to a file. Default is to write to stdout.")
+	fmt.Println("      --C4          Uses C4 as middle C (Yamaha convention). Defaults to C3.")
+	fmt.Println("      --debug       Displays internal information while processing a MIDI file. Defaults to false")
+	fmt.Println("      --verbose     Enables 'verbose' logging. Defaults to false")
 	fmt.Println()
 	fmt.Println("    Example:")
 	fmt.Println()
-	fmt.Println("      midiasm notes --debug --verbose --out one-time.json one-time.mid")
+	fmt.Println("      midiasm export --debug --verbose --out one-time.json one-time.mid")
 	fmt.Println()
 }
 
-func (x Export) Execute(flagset *flag.FlagSet) error {
+func (x export) Execute(flagset *flag.FlagSet) error {
 	filename := flagset.Arg(0)
 
 	smf, err := decode(filename)
@@ -63,8 +62,8 @@ func (x Export) Execute(flagset *flag.FlagSet) error {
 	return x.execute(smf)
 }
 
-func (x Export) execute(smf *midi.SMF) error {
-	op, err := export.NewExport()
+func (x export) execute(smf *midi.SMF) error {
+	op, err := impl.NewExport()
 	if err != nil {
 		return err
 	}
@@ -72,7 +71,7 @@ func (x Export) execute(smf *midi.SMF) error {
 	return x.write(op, smf)
 }
 
-func (x Export) write(op *export.Export, smf *midi.SMF) error {
+func (x export) write(op *impl.Export, smf *midi.SMF) error {
 	out := os.Stdout
 
 	if x.out != "" {